Moved Registration components into new Auth/Registration directory to keep consistency with existing folder structure
