+++ doggii-frontend/src/pages/Auth/Registration/FormShelter.jsx
@@ -1,25 +1,25 @@
-import { useState} from "react";
+import { useState } from "react";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
-import { loginSuccess } from "../redux/slices/authSlice";
-import { setUserInfo } from "../redux/slices/userSlice";
-import Button from "../components/Button";
+import { loginSuccess } from "../../../redux/slices/authSlice";
+import { setUserInfo } from "../../../redux/slices/userSlice";
+import Button from "../../../components/Button";
 
-const FormAdopter = () => {
+const FormShelter = () => {
   const [formData, setFormData] = useState({
-    userName: "",    // Nombre completo
-    email: "",     // Correo electrónico
-    password: "", // Contraseña
-    confirmPassword: "",  // Repetir contraseña
-    termsAccepted: false, // Terminos y condiciones
+    shelterName: "",       // Nombre del refugio
+    userName: "",          // Nombre del responsable
+    email: "",             // Correo electrónico
+    password: "",          // Contraseña
+    confirmPassword: "",   // Repetir contraseña
+    termsAccepted: false,  // Aceptación de términos y condiciones
   });
 
   const [errors, setErrors] = useState({});
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
-  // Función para manejar cambios en los campos del formulario
   const handleChange = (e) => {
     const { name, value, type, checked } = e.target;
 
@@ -27,7 +27,8 @@ const FormAdopter = () => {
       ...formData,
       [name]: type === "checkbox" ? checked : value,
     });
-        // Limpiar el error asociado al campo que se está modificando
+
+    // Limpiar el error asociado al campo que se está modificando
     setErrors((prevErrors) => ({
       ...prevErrors,
       [name]: "", // Limpiar el mensaje de error para este campo
@@ -36,8 +37,9 @@ const FormAdopter = () => {
 
   const validateForm = () => {
     let newErrors = {};
-    if (!formData.userName.trim()) newErrors.userName = "El nombre es obligatorio.";
-    if (!formData.email.trim()) newErrors.email = "El correo es obligatorio.";
+    if (!formData.shelterName.trim()) newErrors.shelterName = "El nombre del refugio es obligatorio.";
+    if (!formData.userName.trim()) newErrors.userName = "El nombre del responsable es obligatorio.";
+    if (!formData.email.trim()) newErrors.email = "El correo electrónico es obligatorio.";
     if (!formData.password) newErrors.password = "La contraseña es obligatoria.";
     if (formData.password !== formData.confirmPassword)
       newErrors.confirmPassword = "Las contraseñas no coinciden.";
@@ -46,7 +48,7 @@ const FormAdopter = () => {
 
     setErrors(newErrors);
     return Object.keys(newErrors).length === 0;
-  }
+  };
 
   const handleSubmit = async (e) => {
     e.preventDefault();
@@ -56,20 +58,23 @@ const FormAdopter = () => {
       return;
     }
 
+    // Validar el formulario antes de enviarlo
     if (!validateForm()) return;
 
+    // Crear el objeto newUser con los datos del formulario
     const newUser = {
       userName: formData.userName,
-      email: formData.email,
       password: formData.password,
+      email: formData.email,
+      shelterName: formData.shelterName,
       admin: false,
-      role: "Adopter",
+      role: "Shelter",
     };
 
     try {
       // Registrar al usuario
       const registerResponse = await axios.post("http://localhost:8082/api/auth/register", newUser);
-
+      
       if (registerResponse.data.success) {
         // Iniciar sesión automáticamente
         const loginResponse = await axios.post("http://localhost:8082/api/auth/login", {
@@ -94,7 +99,7 @@ const FormAdopter = () => {
         }
       }
     } catch (error) {
-      console.error("Error en el registro:", error);
+      console.error("Error al registrar:", error);
       alert("Hubo un error en el registro, intenta nuevamente.");
     }
   };
@@ -109,22 +114,41 @@ const FormAdopter = () => {
           ← Regresar
         </button>
         <img src="/src/assets/logo/inline-logo.png" alt="Logo" className="w-50 h-auto mx-auto mb-4"/>
-        <h1 className="text-2xl font-bold mb-4 text-center">Registro para adoptar</h1>
+        <h1 className="text-2xl font-bold mb-4 text-center">Registro para refugios</h1>
         <form onSubmit={handleSubmit} className="space-y-4">
+          {/* Campo: Nombre del refugio */}
+          <div className='space-y-1'>
+            <label htmlFor='shelterName' className='block text-sm font-medium'>
+              Nombre del refugio
+            </label>
+            <input
+              type="text"
+              name="shelterName"
+              placeholder="Nombre del refugio"
+              value={formData.shelterName}
+              onChange={handleChange}
+              className="w-full px-3 py-2 text-base font-normal rounded-md border border-gray-300 focus:outline-none focus:ring-3 focus:ring-(--secondary-dark) focus:border-white"
+            />
+            {errors.shelterName && <p className="text-red-500 text-sm">{errors.shelterName}</p>}
+          </div>
+
+          {/* Campo: Nombre del responsable */}
           <div className='space-y-1'>
             <label htmlFor='userName' className='block text-sm font-medium'>
-              Nombre completo
+              Nombre del responsable
             </label>
             <input
               type="text"
               name="userName"
-              placeholder="Nombre completo"
+              placeholder="Nombre del responsable"
               value={formData.userName}
               onChange={handleChange}
               className="w-full px-3 py-2 text-base font-normal rounded-md border border-gray-300 focus:outline-none focus:ring-3 focus:ring-(--secondary-dark) focus:border-white"
             />
             {errors.userName && <p className="text-red-500 text-sm">{errors.userName}</p>}
           </div>
+
+          {/* Campo: Correo electrónico */}
           <div className='space-y-1'>
             <label htmlFor='email' className='block text-sm font-medium'>
               Correo electrónico
@@ -139,6 +163,8 @@ const FormAdopter = () => {
             />
             {errors.email && <p className="text-red-500 text-sm">{errors.email}</p>}
           </div>
+
+          {/* Campo: Contraseña */}
           <div className='space-y-1'>
             <label htmlFor='password' className='block text-sm font-medium'>
               Contraseña
@@ -153,6 +179,8 @@ const FormAdopter = () => {
             />
             {errors.password && <p className="text-red-500 text-sm">{errors.password}</p>}
           </div>
+
+          {/* Campo: Repetir contraseña */}
           <div>
             <input
               type="password"
@@ -166,6 +194,8 @@ const FormAdopter = () => {
               <p className="text-red-500 text-sm">{errors.confirmPassword}</p>
             )}
           </div>
+
+          {/* Campo: Aceptar términos y condiciones */}
           <div className="flex items-center gap-2">
             <input
               type="checkbox"
@@ -175,7 +205,7 @@ const FormAdopter = () => {
               className="shrink-0 rounded-md accent-(--primary) checked:bg-(--primary) checked:hover:bg-(--primary) focus:ring-(--primary)"
             />
             <label className="block font-medium">
-              Al crear una cuenta aceptás los{" "}
+              Al crear una cuenta aceptas los{" "}
               <a href="/terms" className="text-(--secondary-dark) hover:text-(--secondary) underline">
                 Términos y condiciones
               </a>
@@ -184,6 +214,8 @@ const FormAdopter = () => {
           {errors.termsAccepted && (
             <p className="text-red-500 text-sm">{errors.termsAccepted}</p>
           )}
+
+          {/* Botón de envío */}
           <Button
             type="submit"
             className="mx-auto text-base"
@@ -196,4 +228,4 @@ const FormAdopter = () => {
   );
 };
 
-export default FormAdopter;
+export default FormShelter;
